Extract facility state patcher
